more IS updates

+++ include/fx.h
@@ -27,7 +27,7 @@ void fx_mesh_save(struct simulation *sim);
 void fx_load_mesh(struct simulation *sim,struct device *in,int number);
 void fx_step();
 int fx_run();
-double fx_get_fx();
+long double fx_get_fx();
 void fx_memory_free();
 int fx_points();
 #endif
